feat: support for 'workspace/configuration'
