+++ src/main/java/com/redhat/devtools/lsp4ij/server/definition/launching/UserDefinedLanguageClient.java
@@ -10,18 +10,11 @@
  ******************************************************************************/
 package com.redhat.devtools.lsp4ij.server.definition.launching;
 
-import com.google.gson.JsonObject;
 import com.intellij.openapi.project.Project;
 import com.redhat.devtools.lsp4ij.LanguageServersRegistry;
 import com.redhat.devtools.lsp4ij.ServerStatus;
 import com.redhat.devtools.lsp4ij.client.LanguageClientImpl;
-import org.eclipse.lsp4j.ConfigurationItem;
-import org.eclipse.lsp4j.ConfigurationParams;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
 /**
  * User defined language client implementation used to track configuration changes
@@ -43,45 +36,6 @@ public class UserDefinedLanguageClient extends LanguageClientImpl {
     public void dispose() {
         super.dispose();
         LanguageServersRegistry.getInstance().removeLanguageServerDefinitionListener(languageServerStartedListener);
-    }
-
-    @Override
-    public CompletableFuture<List<Object>> configuration(ConfigurationParams params) {
-        return CompletableFuture.supplyAsync(() -> {
-            // See https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workspace_configuration
-            List<Object> settings = new ArrayList<>();
-            for (ConfigurationItem item : params.getItems()) {
-                String section = item.getSection();
-
-                Object result = findSettings(section.split("[.]"));
-                // The result is the configuration setting  or null, according to the spec:
-                //  - If a scope URI is provided the client should return the setting scoped to the provided resource.
-                //  - If the client can’t provide a configuration setting for a given scope then null needs to be present in the returned array.
-                settings.add(result);
-            }
-            return settings;
-        });
-    }
-
-
-    private Object findSettings(String[] sections) {
-        var config = createSettings();
-        if (config instanceof JsonObject json) {
-            return findSettings(sections, json);
-        }
-        return null;
-    }
-
-    private static Object findSettings(String[] sections, JsonObject jsonObject) {
-        JsonObject current = jsonObject;
-        for (String section : sections) {
-            Object result = current.get(section);
-            if (!(result instanceof JsonObject json)) {
-                return null;
-            }
-            current = json;
-        }
-        return current;
     }
 
     @Override
